refactor(webview): put every architecture rule on archunit, and empty app/
The suite used archunit to *assert* but selected and counted files with a
`node:fs` walker of its own. Two different answers to "which files are in this
project", and a rule whose population came from one and whose verdict came from
the other can disagree with itself in silence. Selection, counting and verdict
now all go through `projectFiles()`; the shared helper is down to two pure
string predicates and no longer touches disk. `source-files.ts` becomes
`predicates.ts` because that is all it does.

That unlocked the part of the library we had never used: `dependOnFiles`
resolves the real import graph through the TypeScript program. A content regex
cannot follow a `@/` alias to the file it lands on, so it has to trust that the
alias means what its name suggests. Three rules that did not exist before:

- `layer-direction`: app/ -> features/ -> shared/, never upward. The constraint
  the whole restructure rests on, and it had no rule.
- `slice-isolation`: no slice imports a sibling, generated from the `features/`
  listing so a new slice is covered without anyone remembering.
- `no-barrel-files`: the decision the spec argues hardest for, enforced by
  nothing until now.

Each was watched failing against a real violation before being kept, and each
file carries a positive control. Every rule here is a negative, and a negative
passes just as happily when the analyzer resolved nothing at all --
`layer-direction` therefore asserts the app->features edge set is >= 90 (it is
102), so a collapse is reported rather than mistaken for success.

Rule (8) walked *directories* and asked whether each contributed a route. That
caught a folder named `components/` posing as a route segment but was blind to
the real constraint: sixteen components sat loose in folders that did contribute
routes. Rewritten over files, it named all sixteen, and that was the worklist.

Seven were never `app/` code. Placement follows the spec's test -- the type it
renders, not who imports it: header -> user, client-keys-settings -> project,
logs-list -> log, and transaction-stats-table, span-waterfall and
measurements-card -> transaction. theme-selector joins theme-provider in
shared/ui. The remaining nine keep their route and move under `_components/`,
including overview-tiles, which the spec already named as composition no slice
may own.

span-waterfall declared and exported its own `Span`, distinct from the `Span`
agent-trace reads from @rustrak/client: one is a row of the `spans` table, the
other the array embedded in an event payload. It moves to
features/transaction/model/span.ts with the distinction written down, since it
is exactly the case that looks like a reason to reach sideways and is not.

The layout test mocked `'../header'` by relative path. Moving the file does not
make `vi.mock` fail, it makes it not apply -- the real Header loaded and blew up
on `useRouter`. Same silent failure that bit us earlier in this phase.
