+++ apps/webview-ui/src/__tests__/architecture/portable-core.test.ts
@@ -1,34 +1,34 @@
+import { projectFiles } from 'archunit';
 import { describe, expect, it } from 'vitest';
-import {
-  read,
-  rel,
-  SRC,
-  sourceFilesUnder,
-  withoutComments,
-} from './source-files';
+import { isTestFile, withoutComments } from './predicates';
 
 /**
  * The rule this whole phase exists to buy: **the domain does not know Next.**
  *
  * `features/*​/model`, `features/*​/lib` and `shared/lib` hold the logic that is
- * expensive to rewrite and cheap to keep — status derivation, stack-trace
+ * expensive to rewrite and cheap to keep -- status derivation, stack-trace
  * formatting, session-health thresholds, error copy. None of it should care
  * which framework renders it, and a migration off Next should leave all of it
  * untouched.
  *
  * Measured before this phase: 90 of 189 files already imported nothing from
  * `next/*`, and another 44 touched only `next/link` and `next/navigation`,
- * which any framework replaces with a shim. That 71% was an accident of how
- * the code happened to be written. This rule is what turns it into a property.
+ * which any framework replaces with a shim. That 71% was an accident of how the
+ * code happened to be written. This rule is what turns it into a property.
  *
  * Deliberately **not** applied to `ui` segments or to `app/`. Components use
- * `useRouter` and `Link`, and pages are the coupled edge by design — the spec
+ * `useRouter` and `Link`, and pages are the coupled edge by design -- the spec
  * measured the alternative (moving reads to the browser) and rejected it,
  * because the server serves `allow_any_origin()` without credentials and the
  * session cookie would never arrive.
+ *
+ * **Why a content predicate and not `dependOnFiles`.** archunit's graph rules
+ * describe edges between files *in this project*; `next` is an external
+ * package, so there is no node to point at. `adhereTo` is the API for the
+ * remaining case, and it is the one the library documents for exactly this.
  */
 
-const NEXT_IMPORT = /from\s+['"]next(\/|['"])/;
+const NEXT_IMPORT = /(?:from|import)\s+['"]next(\/|['"])/;
 
 const CORE = [
   /(^|\/)features\/[^/]+\/model\//,
@@ -36,23 +36,48 @@ const CORE = [
   /(^|\/)shared\/lib\//,
 ];
 
-const core = sourceFilesUnder(SRC)
-  .filter((f) => !f.includes('__tests__'))
-  .filter((f) => CORE.some((p) => p.test(rel(f))));
+const posix = (path: string) => path.split('\\').join('/');
+const isCore = (path: string) => CORE.some((p) => p.test(posix(path)));
 
 describe('the portable core does not import Next', () => {
-  it('reads the population it expects to read', () => {
-    // 14 files after the phase-6 migration. If this ever collapses toward zero
-    // the globs stopped matching and every file below passes for free, which is
-    // the vacuous-rule failure AD-9 exists to prevent.
+  /**
+   * The floor, and the failure it exists to catch.
+   *
+   * Every assertion here is "no file does X". If the three globs above ever
+   * stop matching -- a segment renamed, a slice restructured -- the rule checks
+   * an empty set and reports success, which is the vacuous-rule failure AD-9
+   * exists to prevent. archunit's own `EmptyTestViolation` does not help: the
+   * filter is `src/**`, which always matches, and the narrowing happens inside
+   * the predicate where the library cannot see it.
+   */
+  it('reads the population it expects to read', async () => {
+    const core = await projectFiles()
+      .inFolder('src/**')
+      .shouldNot()
+      .adhereTo(
+        (file) => !isTestFile(file.path) && isCore(file.path),
+        'counted',
+      )
+      .check();
+
+    // 14 files after the phase-6 migration.
     expect(core.length).toBeGreaterThanOrEqual(14);
   });
 
-  it('has no file importing from next/*', () => {
-    const violations = core
-      .filter((f) => NEXT_IMPORT.test(withoutComments(read(f))))
-      .map(rel);
+  it('has no file importing from next/*', async () => {
+    const rule = projectFiles()
+      .inFolder('src/**')
+      .shouldNot()
+      .adhereTo(
+        (file) =>
+          !isTestFile(file.path) &&
+          isCore(file.path) &&
+          // Comments are blanked first, so the paragraph above -- which names
+          // `next/link` while explaining the rule -- does not trip it.
+          NEXT_IMPORT.test(withoutComments(file.content)),
+        'imports from next/* inside the portable core, which is the coupling this phase removed',
+      );
 
-    expect(violations).toEqual([]);
+    await expect(rule).toPassAsync();
   });
 });
