+++ apps/webview-ui/src/app/(main)/projects/[id]/page.tsx
@@ -4,7 +4,7 @@ import { getProject } from '@/features/project/api/queries';
 import { ProjectHeader } from '@/features/project/ui/project-header';
 import { parseOverviewPeriod } from '@/features/release/model/session-health';
 import { LoadFailure } from '@/shared/ui/load-failure';
-import { OverviewPeriodFilter } from './overview-period-filter';
+import { OverviewPeriodFilter } from './_components/overview-period-filter';
 import {
   CounterTiles,
   CrashFreeTile,
@@ -13,7 +13,7 @@ import {
   SessionHealthTile,
   TileSkeleton,
   TopIssuesTile,
-} from './overview-tiles';
+} from './_components/overview-tiles';
 
 interface ProjectPageProps {
   params: Promise<{ id: string }>;
