+++ apps/webview-ui/src/app/(main)/projects/[id]/performance/page.tsx
@@ -2,8 +2,8 @@ import { Zap } from 'lucide-react';
 import type { Metadata } from 'next';
 import { getProject } from '@/features/project/api/queries';
 import { getTransactionStats } from '@/features/transaction/api/queries';
+import { TransactionStatsTable } from '@/features/transaction/ui/transaction-stats-table';
 import { LoadFailure } from '@/shared/ui/load-failure';
-import { TransactionStatsTable } from './transaction-stats-table';
 
 interface PerformancePageProps {
   params: Promise<{ id: string }>;
