+++ apps/webview-ui/src/app/(main)/settings/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next';
 import { redirect } from 'next/navigation';
 import { getCurrentUser } from '@/features/user/api/queries';
 import { ServiceUnavailable } from '@/shared/ui/service-unavailable';
-import { SettingsMobileNav } from './settings-mobile-nav';
-import { SettingsNav } from './settings-nav';
+import { SettingsMobileNav } from './_components/settings-mobile-nav';
+import { SettingsNav } from './_components/settings-nav';
 
 export const metadata: Metadata = {
   title: 'Settings | Rustrak',
