+++ apps/webview-ui/src/app/invite/[token]/_components/accept-invitation-form.tsx
