+++ apps/webview-ui/src/features/transaction/ui/transaction-stats-table.tsx
